move recorder and context into test_visibility namespace
