+++ sig/datadog/ci.rbs
@@ -30,6 +30,6 @@ module Datadog
 
     def self.components: () -> Datadog::CI::Configuration::Components
 
-    def self.recorder: () -> Datadog::CI::Recorder
+    def self.recorder: () -> Datadog::CI::TestVisibility::Recorder
   end
 end
