+++ sig/datadog/ci/configuration/components.rbs
@@ -2,9 +2,9 @@ module Datadog
   module CI
     module Configuration
       module Components : Datadog::Core::Configuration::Components
-        @ci_recorder: Datadog::CI::Recorder
+        @ci_recorder: Datadog::CI::TestVisibility::Recorder
 
-        attr_reader ci_recorder: Datadog::CI::Recorder
+        attr_reader ci_recorder: Datadog::CI::TestVisibility::Recorder
 
         def initialize: (untyped settings) -> void
