+++ spec/datadog/ci/test_visibility/serializers/factories/test_level_spec.rb
@@ -1,6 +1,6 @@
 require_relative "../../../../../../lib/datadog/ci/test_visibility/serializers/factories/test_level"
 require_relative "../../../../../../lib/datadog/ci/test_visibility/serializers/test_v1"
-require_relative "../../../../../../lib/datadog/ci/recorder"
+require_relative "../../../../../../lib/datadog/ci/test_visibility/recorder"
 
 RSpec.describe Datadog::CI::TestVisibility::Serializers::Factories::TestLevel do
   include_context "CI mode activated" do
